requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bandit==1.7.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
databases==0.8.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
mock==5.1.0
packaging==23.2
pbr==6.0.0
pluggy==1.4.0
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==1.7.1
pytest==8.0.0
pytest-cov==4.1.0
python-decouple==3.8
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
stevedore==5.1.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
